# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: 后台服务接口
  version: 1.0.0
  title: BFFast快速开发平台
basePath: /fast
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/user/list/page:
    get:
      tags:
        - 系统管理-用户管理
      summary: 分页列表
      produces:
        - application/json
      parameters:
        - name: current
          description: 页码
          in: query
          type: integer
          required: true
        - name: size
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: username
          description: 登录账户
          in: query
          type: string
          required: false
        - name: name
          description: 用户姓名
          in: query
          type: string
          required: false
        - name: phone
          description: 手机号
          in: query
          type: string
          required: false
        - name: roleId
          description: 角色id
          in: query
          type: string
          required: false
        - name: lockFlag
          description: 状态 0-正常，9-锁定
          in: query
          type: string
          required: false
        - name: projectId
          description: 项目ID
          in: query
          type: string
          required: false
      responses:
        '200':
          description: 返回用户列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
  /sys/user/list:
    get:
      tags:
        - 系统管理-用户管理
      summary: 不分页列表
      produces:
        - application/json
      parameters:
        - name: name
          description: 用户姓名
          in: query
          type: string
          required: false
        - name: phone
          description: 手机号
          in: query
          type: string
          required: false
        - name: lockFlag
          description: 状态 0-正常，9-锁定
          in: query
          type: string
          required: false
      responses:
        '200':
          description: 返回用户列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
...
  /sys/user/info/{userId}:
    get:
      tags:
        - 系统管理-用户管理
      summary: 获取用户信息
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
      responses:
        200:
          description: 返回用户信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysUserEntity'
  /sys/user/password:
    post:
      tags:
        - 系统管理-用户管理
      summary: 修改密码
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户对象
          in: body
          schema:
            $ref: '#/definitions/PasswordForm'
          required: true
      responses:
        200:
          description: success
          schema:
            $ref: '#/definitions/ResultDTO'
  /sys/user/save:
    post:
      tags:
        - 系统管理-用户管理
      summary: 添加用户
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户对象
          in: body
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/ResultDTO'
  /sys/user/update:
    post:
      tags:
        - 系统管理-用户管理
      summary: 修改用户
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户对象
          in: body
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/ResultDTO'
  /sys/user/delete:
    post:
      tags:
        - 系统管理-用户管理
      summary: 删除用户
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户ID列表
          in: body
          schema:
            $ref: '#/definitions/CommonDeleteDTO'
          required: true
      responses:
        200:
          description: success
          schema:
            $ref: '#/definitions/ResultDTO'
  /sys/user/reset/{userId}:
    post:
      tags:
        - 系统管理-用户管理
      summary: 重置密码
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
      responses:
        200:
          description: success
          schema:
            $ref: '#/definitions/ResultDTO'
  #三方用户-列表-分页
  /user/social/list/page:
    get:
      tags:
        - 系统管理-三方用户
      summary: 分页-列表
      produces:
        - application/json
      parameters:
        - name: current
          description: 页码
          in: query
          type: integer
          required: true
        - name: size
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: username
          description: 帐号
          in: query
          type: string
          required: false
        - name: mobile
          description: 联系方式
          in: query
          type: string
          required: false
        - name: eqpNumber
          description: 设备数量
          in: query
          type: string
          required: false
        - name: type
          description: 注册方式
          in: query
          type: string
          required: false
        - name: startTime
          description: 日期查询-开始时间
          in: query
          type: string
          required: false
        - name: endTime
          description: 日期查询-结束时间
          in: query
          type: string
          required: false
      responses:
        '200':
          description: 返回三方用户列表
          schema:
            $ref: '#/definitions/SysUserSocialEntityListPage'
  #三方用户-列表-不分页
...
  /user/social/info/{id}:
    get:
      tags:
        - 系统管理-三方用户
      summary: 单条
      produces:
        - application/json
      parameters:
        - name: id
          description: 用户ID
          in: path
          type: integer
          required: true
      responses:
        200:
          description: 返回信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysUserSocialEntity'
  #三方用户-列表-分页
...
  /sys/user/open/list/page:
    get:
      tags:
        - 系统管理-开放平台
      summary: 账号列表-分页
      produces:
        - application/json
      parameters:
        - name: current
          description: 页码
          in: query
          type: integer
          required: true
        - name: size
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: companyName
          description: 企业名称
          in: query
          type: string
          required: false
        - name: openKey
          description: openKey
          in: query
          type: string
          required: false
        - name: isLock
          description: 是否锁定 0-否 1-是
          in: query
          type: string
          required: false
      responses:
        200:
          description: 返回列表
          schema:
            $ref: '#/definitions/SysUserOpenEntityListPage'

  #系统管理-开放平台-列表-不分页
...
  /sys/user/open/save:
    post:
      tags:
        - 系统管理-开放平台
      summary: 保存账号
      produces:
        - application/json
      parameters:
        - name: body
          description: SysUserOpen 对象
          in: body
          schema:
            $ref: '#/definitions/SysUserOpenEntity'
          required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/ResultDTO'

  #系统管理-开放平台-修改
  /sys/user/open/update:
    post:
      tags:
        - 系统管理-开放平台
      summary: 修改账号
      produces:
        - application/json
      parameters:
        - name: body
          description: SysUserOpen 对象
          in: body
          schema:
            $ref: '#/definitions/SysUserOpenEntity'
          required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/ResultDTO'

  #系统管理-开放平台-单条查询
  /sys/user/open/info/{id}:
    get:
      tags:
        - 系统管理-开放平台
      summary: 查询单条账号
      produces:
        - application/json
      parameters:
        - name: id
          description: 记录ID
          in: path
          type: integer
          required: true
      responses:
        200:
          description: 返回单条记录
          schema:
            type: object
            properties:
              code:
                description: 状态码 0-成功  非0-失败
                type: integer
                format: int32
              msg:
                description: 返回结果消息
                type: string
              data:
                $ref: '#/definitions/SysUserOpenEntity'
  #系统日志-开放接口三方调用记录-列表-分页
...
definitions:
  ResultDTO:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
        default: 0
      msg:
        description: 失败原因
        type: string
        default: success
      data:
        description: 数据对象
        type: object
  CommonDeleteDTO:
    type: array
    items:
      type: integer
      format: int64
...
  PasswordForm:
    type: object
    properties:
      password:
        description: 原密码
        type: string
      newPassword:
        description: 新密码
        type: string
  SysUserEntity:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      name:
        description: 用户姓名
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      lockFlag:
        description: 状态 0-正常，9-锁定
        type: string
      imageUrl:
        description: 头像
        type: string
      lastTime:
        description: 上次登录时间
        type: integer
        format: int64
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int64
      areaName:
        description: 地区名称
        type: string
      createUserId:
        description: 创建者ID
        type: integer
        format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
  SysUserEntityList:
    type: object
    properties:
      code:
        description: 状态码 0-成功  非0-失败
        type: integer
        format: int32
      msg:
        description: 返回结果消息
        type: string
      data:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserEntity'
  SysUserEntityEdit:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      password:
        description: 密码
        type: string
      name:
        description: 用户姓名
        type: string
      phone:
        description: 手机号
        type: string
      roleId:
        description: 角色id
        type: integer
        format: int64
      roleCode:
        description: 角色code
        type: string
      linkMan:
        description: 联系人
        type: string
      linkPhone:
        description: 联系方式
        type: string
      area:
        description: 地区
        type: string
      mapX:
        description: 地图x轴
        type: string
      mapY:
        description: 地图y轴
        type: string
      areaMap:
        description: 地图地址
        type: string
      projectId:
        description: 项目id
        type: integer
        format: int64
      agentId:
        description: 项目id
        type: integer
        format: int64
      lockFlag:
        description: 状态 0-正常，9-锁定
        type: string
      imageUrl:
        description: 头像
        type: string
...
  SysUserSocialEntityListPage:
    type: object
    properties:
      code:
        description: 状态码 0-成功  非0-失败
        type: integer
        format: int32
      msg:
        description: 返回结果消息
        type: string
      data:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserSocialDTO'
...
  SysUserSocialEntity:
    type: object
    properties:
      id:
        description: id
        type: integer
        format: int64
      providerId:
        description: 注册方式
        type: integer
        format: int64
      providerUserId:
        description: 注册用户ID
        type: integer
        format: int64
      rank:
        description: rank
        type: string
      eqpNumber:
        description: eqpNumber
        type: string
      displayName:
        description: 昵称
        type: string
      profileUrl:
        description: profileUrl
        type: string
      imageUrl:
        description: imageUrl
        type: string
      accessToken:
        description: accessToken
        type: string
      secret:
        description: secret
        type: string
      refreshToken:
        description: refreshToken
        type: string
      expireTime:
        description: expireTime
        type: string
  #三方用户 entity
  SysUserSocialDTO:
    type: object
    properties:
      userId:
        description: userId
        type: integer
        format: int64
      imageUrl:
        description: 头像
        type: integer
        format: int64
      name:
        description: 昵称
        type: integer
        format: int64
      mobile:
        description: 联系方式
        type: string
      username:
        description: 帐号
        type: string
      eqpNumber:
        description: 设备数量
        type: string
      type:
        description: 注册方式
        type: string
      lockFlag:
        description:  0-正常，9-锁定
        type: string
      lastTime:
        description:  上次登录时间
        type: string
      createTime:
        description:  注册时间
        type: string
  #系统管理-开放平台-EntityListPage
  SysUserOpenEntityListPage:
    type: object
    properties:
      code:
        description: 状态码 0-成功  非0-失败
        type: integer
        format: int32
      msg:
        description: 返回结果消息
        type: string
      data:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserOpenEntity'
  #系统管理-开放平台-EntityList
...
  SysUserOpenEntity:
    type: object
    properties:
      id:
        description: 主键id
        type: integer
        format: int64
      openKey:
        description: 开放账户key
        type: string
      openSecret:
        description: 开放账户secret
        type: string
      openIp:
        description: 开放账户id集
        type: string
      companyName:
        description: 公司名称
        type: string
      contact:
        description: 联系人
        type: string
      contactNumber:
        description: 联系电话
        type: string
      businessMan:
        description: 商务对接人
        type: string
      isLock:
        description: 是否锁定 0-否 1-是
        type: string
      isDel:
        description: 是否删除 0-否 1-是
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date
  #系统日志-开放接口三方调用记录-EntityListPage
...